refactor: Move commands related to Settings App to the separate module
