+++ lib/tools/index.js
@@ -5,6 +5,7 @@ import apkSigningMethods from './apk-signing.js';
 import apkUtilsMethods from './apk-utils.js';
 import apksUtilsMethods from './apks-utils.js';
 import emuMethods from './adb-emu-commands.js';
+import settingsClientCommands from './settings-client-commands';
 
 Object.assign(
     methods,
@@ -14,6 +15,7 @@ Object.assign(
     apkSigningMethods,
     apkUtilsMethods,
     apksUtilsMethods,
+    settingsClientCommands,
 );
 
 export default methods;
